Move graph-node non assemblyscript code to util package
