+++ packages/graph-node/src/index.ts
@@ -1,2 +1,1 @@
-export * from './state-utils';
 export * from './watcher';
