--- packages/graph-node/src/types.ts
@@ -1,35 +1,0 @@
-//
-// Copyright 2021 Vulcanize, Inc.
-//
-
-// Enum types from @graphprotocol/graph-ts.
-
-export enum EthereumValueKind {
-  ADDRESS = 0,
-  FIXED_BYTES = 1,
-  BYTES = 2,
-  INT = 3,
-  UINT = 4,
-  BOOL = 5,
-  STRING = 6,
-  FIXED_ARRAY = 7,
-  ARRAY = 8,
-  TUPLE = 9,
-}
-
-export enum Level {
-  CRITICAL = 0,
-  ERROR = 1,
-  WARNING = 2,
-  INFO = 3,
-  DEBUG = 4,
-}
-
-export enum JSONValueKind {
-  NULL = 0,
-  BOOL = 1,
-  NUMBER = 2,
-  STRING = 3,
-  ARRAY = 4,
-  OBJECT = 5,
-}
