+++ packages/util/src/graph/types.ts
@@ -59,6 +59,19 @@ export enum TypeId {
   ArrayBigDecimal = 51,
 }
 
+export enum EthereumValueKind {
+  ADDRESS = 0,
+  FIXED_BYTES = 1,
+  BYTES = 2,
+  INT = 3,
+  UINT = 4,
+  BOOL = 5,
+  STRING = 6,
+  FIXED_ARRAY = 7,
+  ARRAY = 8,
+  TUPLE = 9,
+}
+
 export enum ValueKind {
   STRING = 0,
   INT = 1,
@@ -68,4 +81,21 @@ export enum ValueKind {
   NULL = 5,
   BYTES = 6,
   BIGINT = 7,
+}
+
+export enum Level {
+  CRITICAL = 0,
+  ERROR = 1,
+  WARNING = 2,
+  INFO = 3,
+  DEBUG = 4,
+}
+
+export enum JSONValueKind {
+  NULL = 0,
+  BOOL = 1,
+  NUMBER = 2,
+  STRING = 3,
+  ARRAY = 4,
+  OBJECT = 5,
 }
