+++ packages/util/src/index.ts
@@ -21,4 +21,5 @@ export * from './server';
 export * from './graph/graph-decimal';
 export * from './graph/database';
 export * from './graph/utils';
+export * from './graph/state-utils';
 export * from './graph/types';
